feat: 게임 시작 초기 화면 구현
