+++ TtouchIsland/Sources/AppModel.swift
@@ -9,15 +9,15 @@ class AppModel {
     var gameAudioRoot: Entity? {
         gameRoot?.children.first(where: { $0.name == "Root" })
     }
-
+    
     var displayOverlaysVisible = false
     var collectedCoin = false
     var collectedKey = false
     let isPortrait = true
     var levelFinished = false
-
+    
     var metalDevice: MTLDevice? = MTLCreateSystemDefaultDevice()
-
+    
     func reset() {
         gameRoot?.removeFromParent()
         gameRoot = nil
